requirements.txt
python-dotenv>=1.0.1
google-generativeai>=0.7.2
langchain>=0.3.7
langchain-core>=0.3.15
langchain-google-genai>=0.3.4